requirements.txt
# python >= 3.10
swarm
reportlab
google-generativeai
python-dotenv
speechrecognition
pyaudio
